Move examples to separate files
